+++ src/webpack-config.js
@@ -12,8 +12,8 @@ const config = {
     global: false,
     process: false,
     Buffer: false,
-    __filename: false,
-    __dirname: false
+    __filename: true,
+    __dirname: true
   },
 
   plugins: [